add modifications to controller, models and properties

+++ src/main/java/unah/lenguajes1700/josemartinez/examen2/Models/Cuotas.java
@@ -18,8 +18,8 @@ public class Cuotas {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long codigoCuota ;
-    private Long mes;
+    private Integer codigoCuota ;
+    private Integer mes;
 
     @JsonIgnore
     @ManyToOne
